added docs, changed connect for server part
