+++ core/scripts/server.ts
@@ -23,7 +23,6 @@ const isProd = process.env.NODE_ENV === 'production'
 process['noDeprecation'] = true
 
 const app = express()
-require('express-ws')(app)
 
 serverExtensions.serverModules.forEach(serverModule => {
   if (Array.isArray(serverModule)) {
@@ -135,43 +134,6 @@ app.use('/service-worker.js', serve('dist/service-worker.js', false, {
 
 app.post('/invalidate', invalidateCache)
 app.get('/invalidate', invalidateCache)
-
-if (config.clearCache && config.clearCache.enabled) {
-  let wsClients = []
-
-  if (config.clearCache.websocket.enabled) {
-    app.ws(config.clearCache.websocket.endpoint, (ws, req) => {
-      const index = wsClients.push(ws) - 1
-
-      ws.on('message', msg => {
-        if (msg === config.clearCache.websocket.key) {
-          for (let client of wsClients) {
-            client.send(msg)
-          }
-        }
-      })
-
-      ws.on('open', () => {
-        console.log('Webscoket - connection established')
-      })
-
-      ws.on('error', err => {
-        console.error('Websocket error - ', err)
-      })
-
-      ws.on('close', () => {
-        wsClients.splice(index, 1)
-      })
-    })
-  }
-
-  if (config.clearCache.version.enabled) {
-    app.get(config.clearCache.version.endpoint, (req, res) => {
-      const version = process.env[config.clearCache.version.envName] || 1
-      apiStatus(res, version, 200)
-    })
-  }
-}
 
 app.get('*', (req, res, next) => {
   if (NOT_ALLOWED_SSR_EXTENSIONS_REGEX.test(req.url)) {
